Change precision's type from double to float

+++ bbp/parallel/include/bbp_parallel.cpp
@@ -1,8 +1,8 @@
 #include "bbp_parallel.h"
 
-mila::bbp::parallel::BBP::BBP(): precision_(1e-5) {
+mila::bbp::parallel::BBP::BBP() : precision_(1e-5f) {
 }
 
-double mila::bbp::parallel::BBP::precision() const {
+float mila::bbp::parallel::BBP::precision() const {
   return precision_;
 }
